wirc 1.1.3
wirc 1.1.3
CRITICAL Fix to the color protocol that wasnt working properly because of a missing +1 in a positionting calculation. Not much appart of that but i wanted to put on a working version before returning to the normal life cycle.

wirc 1.1.2
Again, this update brings a lot of bug fixes and new features. Unfortunately, I didn't take any notes so here are the highlights.

Added scripting support for on_load
       /scripts/on_load will be executed at program start, we could use this gate to load more script but no event or variable is supported right now.
Added full color scheme
      All colors are stored in wirc.ini but can now be changed using the Application Schema Editor. wIRC UI uses 16 colors for visual customization which are
      editable and mirc CTRL+K 16 sub-protocol color model can now also be edited. Added additional color for CTRL+K (16 point to wircForeOWN and
      17 point to wircBackGROUND)
Improved mIRC Color sub-protocol
     speed improvement, limiting the 'colortext' field to 525 bytes so that it can process an entire irc packet
Fixed channel modes +k and +l are not set correctly
     The user must always know the key (+k) to be able to delete it because we do not store any type of password
Many more
     Im not much of a Doc guy. but i will keep on with the big lines.

This update is the last "quick update" as my days off are coming to an end. Development of wIRC is still ongoing, but at a slower pace.